# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SocialConnect API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterRequest {
  email: string;
  /**
   * @minLength 3
   * @maxLength 30
   * @pattern ^[a-zA-Z0-9_]+$
   */
  username: string;
  /** @minLength 8 */
  password: string;
  first_name: string;
  last_name: string;
}

export interface LoginRequest {
  /** Email or username */
  login: string;
  password: string;
}

export interface UserProfile {
  id: number;
  username: string;
  email: string;
  first_name: string;
  last_name: string;
  bio?: string | null;
  avatar_url?: string | null;
  website?: string | null;
  location?: string | null;
  posts_count: number;
  followers_count: number;
  following_count: number;
  /** Whether the current user follows this user */
  is_following: boolean;
  created_at: string;
  last_login?: string | null;
}

export interface AuthResponse {
  access_token: string;
  user: UserProfile;
}

export interface UserListResponse {
  users: UserProfile[];
  total: number;
  page: number;
  limit: number;
  has_next: boolean;
}

export interface UpdateProfileRequest {
  /** @maxLength 160 */
  bio?: string | null;
  avatar_url?: string | null;
  website?: string | null;
  location?: string | null;
  first_name?: string;
  last_name?: string;
}

export interface AvatarUploadResponse {
  avatar_url: string;
  message: string;
}

export interface PostAuthor {
  id: number;
  username: string;
  first_name: string;
  last_name: string;
  avatar_url?: string | null;
}

export interface Post {
  id: number;
  content: string;
  image_url?: string | null;
  is_active: boolean;
  like_count: number;
  comment_count: number;
  /** Whether the current user liked this post */
  is_liked: boolean;
  author: PostAuthor;
  created_at: string;
  updated_at: string;
}

export interface PostListResponse {
  posts: Post[];
  total: number;
  page: number;
  limit: number;
  has_next: boolean;
}

export interface UpdatePostRequest {
  /** @maxLength 280 */
  content?: string;
}

export interface Comment {
  id: number;
  content: string;
  author: PostAuthor;
  post_id: number;
  created_at: string;
  updated_at: string;
}

export interface AddCommentRequest {
  /**
   * @minLength 1
   * @maxLength 500
   */
  content: string;
}

export interface CommentListResponse {
  comments: Comment[];
  total: number;
  page: number;
  limit: number;
  has_next: boolean;
}

export type ListUsersParams = {
  /**
   * @minimum 1
   */
  page?: number;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
};

export type GetUserFollowersParams = {
  /**
   * @minimum 1
   */
  page?: number;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
};

export type GetUserFollowingParams = {
  /**
   * @minimum 1
   */
  page?: number;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
};

export type UploadAvatarBody = {
  avatar: Blob;
};

export type ListPostsParams = {
  /**
   * @minimum 1
   */
  page?: number;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
};

export type CreatePostBody = {
  /** @maxLength 280 */
  content: string;
  image?: Blob;
};

export type ListCommentsParams = {
  /**
   * @minimum 1
   */
  page?: number;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
};

export type GetFeedParams = {
  /**
   * @minimum 1
   */
  page?: number;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
};