# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SocialConnect API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logoutUser
      tags: [auth]
      summary: Logout user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        username:
          type: string
          minLength: 3
          maxLength: 30
          pattern: "^[a-zA-Z0-9_]+$"
        password:
          type: string
          minLength: 8
        first_name:
          type: string
        last_name:
          type: string
      required:
        - email
        - username
        - password
        - first_name
        - last_name
    LoginRequest:
      type: object
      properties:
        login:
          type: string
          description: Email or username
        password:
          type: string
      required:
        - login
        - password
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - access_token
        - user
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        bio:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        posts_count:
          type: integer
        followers_count:
          type: integer
        following_count:
          type: integer
        is_following:
          type: boolean
          description: Whether the current user follows this user
        created_at:
          type: string
          format: date-time
        last_login:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - username
        - email
        - first_name
        - last_name
        - posts_count
        - followers_count
        - following_count
        - is_following
        - created_at
...